requirement_1.txt
aiofiles==22.1.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.2
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blis==0.7.9
Boruta==0.3
bs4==0.0.1
catalogue==2.0.8
catboost==1.1.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
colorlover==0.3.0
confection==0.0.3
constantly==15.1.0
cryptography==38.0.3
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.28
daphne==4.0.0
dashboard==0.0.6
databases==0.6.2
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
decouple==0.0.7
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.3
docker==6.0.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
funcy==1.17
future==0.18.2
gcloud==0.18.3
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.29
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.9.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
incremental==22.10.0
ipykernel==6.17.0
ipython==8.6.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-widgets==3.0.3
jwcrypto==1.4.2
jws==0.1.3
kiwisolver==1.4.4
kmodes==0.12.2
langcodes==3.3.0
lightgbm==3.3.3
llvmlite==0.39.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
missingno==0.5.1
mlflow==1.30.0
mlxtend==0.21.0
multimethod==1.9
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
numba==0.56.4
numexpr==2.8.4
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
orm==0.2.0.dev1
packaging==21.3
pandas==1.5.1
pandas-profiling==3.4.0
parso==0.8.3
pathy==0.6.2
patsy==0.5.3
phik==0.12.2
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
plac==1.3.5
platformdirs==2.5.3
plotly==5.11.0
preshed==3.0.8
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.32
protobuf==4.21.9
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==2.2.2
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyLDAvis==3.3.1
pynndescent==0.5.8
pyod==1.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==4.0.0
python-multipart==0.0.5
pytz==2022.6
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.9.3
seaborn==0.12.1
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post1
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
sqlparse==0.4.3
srsly==2.4.5
stack-data==0.6.0
starlette==0.21.0
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
textblob==0.17.1
thinc==8.1.5
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typer==0.4.2
typesystem==0.3.0.dev0
typing_extensions==4.4.0
tzdata==2022.6
umap-learn==0.5.3
Unipath==1.1
urllib3==1.26.12
uvicorn==0.19.0
virtualenv==20.16.6
visions==0.7.5
waitress==2.1.2
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wordcloud==1.8.2.2
wrapt==1.14.1
xgboost==1.7.1
yarg==0.1.9
yellowbrick==1.5
zipp==3.10.0
zope.interface==5.5.1
